projects/llm-training/requirements.txt
torch>=2.2.0
numpy>=1.24.0
transformers>=4.40.0
tiktoken>=0.7.0
trl>=0.8.0
mlflow>=2.13.0
pyyaml>=6.0
google-cloud-storage>=2.10.0
tensorboard>=2.15.0
